api: refactor authentication to be passwordless

+++ src/main/java/me/finnthorben/thingpeoplelist/security/auth/AuthServiceImpl.java
@@ -2,55 +2,37 @@ package me.finnthorben.thingpeoplelist.security.auth;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import me.finnthorben.thingpeoplelist.security.sessions.Session;
-import me.finnthorben.thingpeoplelist.security.sessions.SessionRepository;
+import me.finnthorben.thingpeoplelist.security.sessions.*;
 import me.finnthorben.thingpeoplelist.users.User;
-import org.springframework.security.authentication.ReactiveAuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContext;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.support.TransactionTemplate;
 import reactor.core.publisher.Mono;
 import reactor.core.scheduler.Scheduler;
 
-import javax.transaction.Transactional;
+import java.time.ZonedDateTime;
 import java.util.UUID;
 
 @Service
 @Slf4j
 @RequiredArgsConstructor
 public class AuthServiceImpl implements AuthService {
-
-    private final ReactiveAuthenticationManager authManager;
     private final SessionRepository sessionRepository;
+
+    private final PendingSessionRepository pendingSessionRepository;
+
+    private final SessionService sessionService;
 
     private final Scheduler jdbcScheduler;
 
     private final TransactionTemplate transactionTemplate;
 
     @Override
-    public Mono<Session> login(String username, String password, String ipAddress, String userAgent) {
-        return Mono
-                .just(new UsernamePasswordAuthenticationToken(username, password))
-                .flatMap(authManager::authenticate)
-                .doOnNext(auth -> {
-                    // set authentication for the current thread
-                    SecurityContext securityContext = SecurityContextHolder.getContext();
-                    securityContext.setAuthentication(auth);
-                })
-                .map(auth -> {
-                    // construct a session from the validated authentication
-                    Session session = new Session((User) auth.getPrincipal(), ipAddress, userAgent);
-                    return sessionRepository.save(session);
-                })
-                .subscribeOn(jdbcScheduler);
-    }
-
-    @Override
-    public Mono<Void> logout(UUID sessionId) {
+    public Mono<Void> logoutSession(User user, UUID sessionId) throws CantLogoutLastSessionException {
         return Mono.fromRunnable(() -> transactionTemplate.execute(status -> {
+                    if (sessionRepository.findByUser(user).size() == 1) {
+                        throw new CantLogoutLastSessionException();
+                    }
+
                     sessionRepository.deleteById(sessionId);
                     return null;
                 }))
@@ -59,7 +41,7 @@ public class AuthServiceImpl implements AuthService {
     }
 
     @Override
-    public Mono<Void> logout(User user, Session exceptSession) {
+    public Mono<Void> logoutExceptSession(User user, Session exceptSession) {
         return Mono.fromRunnable(() -> transactionTemplate.execute(status -> {
                     sessionRepository.deleteByUserAndIdIsNot(user, exceptSession.getId());
                     return null;
@@ -69,12 +51,26 @@ public class AuthServiceImpl implements AuthService {
     }
 
     @Override
-    public Mono<Void> logout(User user) {
-        return Mono.fromRunnable(() -> transactionTemplate.execute(status -> {
-                    sessionRepository.deleteByUser(user);
-                    return null;
-                }))
-                .cast(Void.class)
+    public Mono<Session> login(String pendingSessionToken, String ipAddress, String userAgent) throws InvalidPendingSessionException {
+        return Mono
+                .fromCallable(() -> pendingSessionRepository.findByToken(pendingSessionToken))
+                .map(pendingSession -> pendingSession.orElseThrow(InvalidPendingSessionException::new))
+                .flatMap(pendingSession -> {
+                    if (!pendingSession.isActive()) {
+                        return Mono.error(new AuthService.InvalidPendingSessionException());
+                    }
+                    return sessionService.createSession(pendingSession.getUser(), ipAddress, userAgent);
+                })
+                .subscribeOn(jdbcScheduler);
+    }
+
+    @Override
+    public Mono<PendingSession> createPendingSession(User user) {
+        return Mono
+                .fromCallable(() -> {
+                    PendingSession session = new PendingSession(user, ZonedDateTime.now().plusMinutes(30));
+                    return pendingSessionRepository.save(session);
+                })
                 .subscribeOn(jdbcScheduler);
     }
 }
